+++ src/containers/App.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import App from '../App';
+import App from '../components/App';
 
 import { initializeEvents } from '../actions';
